refactor(abstract-utxo): consolidate and refactor unit tests
Consolidate test fixtures and refactor unit tests for better organization.
Move address validation tests to their own file, reorganize BTC backup key
fixtures and audit tests, and consolidate spoofing protection tests.

Issue: BTC-2650
