--- modules/abstract-utxo/test/unit/impl/btc/unit/fixtures/index.ts
@@ -1,1 +1,0 @@
-export * from './btcBackupKey';
